[Tools/gsc] More resilient naming options for GSC
Currently, GSC generates two images: signed and non-signed,
whose names are determined automatically by adding the prefix "gsc-" and
the postfix "-unsigned", respectively.
This is less flexible, not very consistent (if you include tag
name), and also could not work in some cases if the image has a repository.
Instead of mixing prefix and postfix scheme, we add postfix to the image
name excluding the tag.
Also, we let users to set the name of the output image with '--name' or '-o' flag.