[Examples] tensorflow: Fix non-existing target 'default'
The commit: [Examples] tensorflow: Fix default `make` target
replaced the target 'default' with 'all'.

However, the targets 'run-graphene', 'run-native' and 'check'
still depend on 'default', so it is not possible to run
  make run-native
  make run-graphene
  make run-graphene SGX=1
as documented in the README.md file.

To fix this, modify these targets to depend on 'all' instead, as the target
'default' was replaced by 'all'.