[LibOS] Prevent reserving inaccessible memory for a checkpoint
Sometimes a checkpoint may reside in parent's process at addresses not
available in the child (e.g. reserved by PAL).
receive_checkpoint_and_restore() had a bug in the allocation logic which
tried to allocate the memory at the same address as in the parent even
if this was not possible, leading to failures (note that
bkeep_mmap_fixed() is a low-level API and allows bookkeeping memory
outside of LibOS range).

The bug was detected when trying to run R example.