[Pal] regression: Remove meaningless AvxDisable test
Previously, this test checked if the test binary fails with
sgx.require_avx = 0 in the manifest (since SGX disallows the enclave to
use AVX and the "Illegal instruction" failure is produced).

In "[Pal/Linux-SGX] Change manifest option sgx.enable_[avx|avx512|mpx]
to…" commit we changed the semantics of sgx.require_avx. If it is unset
(like in AvxDisable), SGX PAL will dynamically decide whether to
use AVX or not, depending on the machine it's running on. So, machines
without AVX will fail with "Illegal instruction", but machines with AVX
will not fail (and proceed to return 1 which is like a benign failure).

So the test works correctly but now it doesn't really test anything (it
always fails, either because AVX instructions are not available or
because it reaches return 1).