[Pal/Linux-SGX] Remove unneeded manifest option `sgx.zero_heap_on_demand`
Previously, we introduced `sgx.zero_heap_on_demand` in Linux-SGX as a
knob to trade off runtime degradation on memory allocations for faster
enclave start-up times. This was an incorrect fix because Linux-SGX's
`_DkVirtualMemoryAlloc()` always zeros the requested memory region,
so there was a double-zero of the heap at runtime. Note that POSIX
requires `_DkVirtualMemoryAlloc()` to zero-out the memory, and many
applications rely on this (Apache, Blender in my experiments). Thus,
this commit keeps the zero-out in `_DkVirtualMemoryAlloc()` and
removes zero-outs on enclave init and in `get_enclave_pages()`.
This renders `sgx.zero_heap_on_demand` useless, so this manifest
option is also removed. Also note that this commit doesn't introduce
any performance degradation (in fact, now Graphene behaves as if
`sgx.zero_heap_on_demand = 1` always).