[LibOS] Mark checkpointed dir handle and dentry as not listed
The "directory" shim_handle and its associated dentry are marked as
"listed" on getdents(), caching the contents of the directory as a
list of dentries. However, Graphene doesn't checkpoint these contents
(unless they were actually opened). Previously, there was a bug of not
cleaning up the "listed" state of dir dentry and associated pointers
of dir shim_handle on fork/clone. This led to getdents() in the child
process showing empty directory contents. This commit fixes this bug.