[GSC] Add an option to pre-build Graphene-only image
Previous iterations of GSC always built the Graphene runtime as part of
the `gsc build` command. This commit adds an option to extract the
compilation of the Graphene runtime from the GSC build command into a
separate command called `gsc build-graphene`. Using `gsc build-graphene`,
one can prepare a Docker image that includes the required runtime files
for a `gsc build` command.

The purpose of this separation is to publish Graphene Docker images for
special environments such as cloud environments. This simplifies the
configuration parameters for the user, since they only have to specify
the Graphene Docker image name instead of the Graphene repository and
Driver details.