[Pal] Rewrite IS_ARRAY macro
The (((void *) &(arg)) == ((void *) (arg))) trick does not work
with clang, because clang does not accept it as a constant
expression and doesn't allow using it in static_assert.

I was not able to write a generic (any type) IS_ARRAY macro
compatible with static_assert, but here we only need to check for
an array of chars.