[Pal/Linux-SGX] Rewrite GDB symbol file updates
See https://github.com/oscarlab/graphene/issues/1814.

The old integration was based on directly executing GDB commands.
Instead, we maintain a linked list of debug sections, and notify
GDB using a breakpointed function.

In addition, make sure to read the structure when we attach to a
new process, by hooking the GDB stop event.

From my testing, this also solves the problem with breakpoints not
working until you break (Ctrl-C) inside an enclave. Adding a
breakpoint to a specific file (`break foo.c:main`) also seems to
work correctly, even before the process is started (the breakpoint
is conditional on a file load).