Create a helper script for running pytest
Running a subset of regression tests, or a single test, is
currently rather annoying, as you cannot simply run 'pytest':
you need to override PYTHONPATH and specify several environment
variables.

Instead, move the whole boilerplate to a wrapper scripts that
allows running pytest with any command line options. Use it in
Makefiles as well, so that it doesn't go out of sync with CI.