[LibOS] test/ltp: Allow skipping LTP tests by fnmatch
Now one can write `*` into section names in ltp.cfg, but such
sections can only contain "skip". This is in order to make the
configuration file more usable: in many cases, we want to skip
a whole class of tests, because a given feature/syscall is not
supported by Graphene.