[test] allow skipping LTP tests by fnmatch
From now you can write * into section names in ltp.cfg, but such
sections can only contain "skip".

This is in order to make the configuration file more usable: in
many cases, we want to skip a whole class of tests, because a given
feature/syscall is not supported by Graphene.