[Examples] Remove python-sh-insecure example
The purpose of this example was to show how to break a cyclic dependency
when spawning processes, using an ugly hack. This hack will stop working
after manifest resolution cleanup, which will be followed by a loader
rework, which finally will make this use-case work out-of-the-box,
without requiring any hacks.