[LibOS] test/ltp: Start tests in order, sort results
Due to how asyncio.gather() works, the tests are started in a
random order. Even if exact order cannot be ensured (because the
tests run in parallel), this makes for a pretty bad user
experience, as you get totally different output every time, and
it's hard to tell how far along is the test run.

Instead, start the tests in the order they are specified in the
runtest file.

In addition, sort the test results in XML report, so that it is
easier to analyze.