[LibOS] Fix dangling pointer `dent` on connect(AF_UNIX)
When connecting on a named UNIX domain socket, the connect()
syscall checks if the socket path exists via path_lookupat(&dent).
However, path_lookupat() may return error and *not* update dent
allocated on stack. In this case, dent contains garbage, and Graphene
fails. This commit fixes this via explicit `dent = NULL`. A `unix`
LibOS test is updated to test this.