[Pal/Linux-SGX] Support in-kernel SGX driver v39
SGX driver v39 removes `count` field and returns "number of bytes added"
as return value of the SGX_IOC_ENCLAVE_ADD_PAGES ioctl. It also caps
the maximum number of bytes to be added as 1MB, or 256 pages. Thus, we
must loop on the ADD_PAGES ioctl until all pages are added and first
check `ret > 0` and only then check `count` field (to support all
versions of the SGX driver).