[LibOS] test/regression: Fix exit_group test
The exit code of the test was a random number between 0 and 4,
which gave no guarantee that the right code from exit() will
actually reach the user.

Instead of testing a race between threads, make it possible to
call exit() in a chosen thread, and verify that it always works.

In addition, propagate the exit code from a forked process. That
use case was broken recently and has been fixed in the commit
titled 'Rework threads implementation'.