[LibOS] Fix type field in dirent struct
Currently in graphene, the type field in dirent struct is only updated
for parent directory and is set to DT_UNKNOWN type for its children.
But APIs like sysconf(_SC_NPROCESSORS_CONF) rely on type field to
identify the number of processors on the host by reading the number of
cpuX directories and ensuring their type is set to DT_DIR. This patch
addresses the issue by updating the dirent type appropriately even for
child directories or files.