[LibOS] Fix fcntl(F_DUPFD) hang
The operation loops indefinitely on error. Instead, it should find
the first free FD, and then try allocating it.

In addition, the right error after exceeding the limit is EMFILE
(however, dup2() is still supposed to return EBADF if asking for
an out-of-range value, as checked by the dup201 LTP test).