[LibOS] Unlock before blocking accept() in shim_do_accept()
The accept() emulation locks the shim handle at the beginning of
shim_do_accept() and then proceeds to issue a blocking accept()
host-level syscall. This blocked lock may lead to deadlocks, e.g.
due to a fork in another thread (fork emulation must checkpoint the
shim handle, for which it needs to grab the lock but it is blocked).
This commit unlocks right before the blocking accept() syscall and
thus avoids the deadlock.