[CI] clean-check: ignore from python/graphenelibos/__pycache__
The python/ subdir does not have makefiles (only meson.build), so there
is no {,dist}clean. The problem stems from the fact that the python
package is imported from the tree as part of SGX tests (specifically,
on sign and get-token), so the __pycache__ is created from makefiles in
other subdirs.

This all will become irrelevant when either of the two happens:
- the tests would be run on installed system -- then the __pycache__
  will be in /usr/lib/python3*); or
- whole graphene will be built by meson -- then the __pycache__ will
  happen inside builddir and there will be no more clean-check.