[Pal/lib] Remove "api.h" include from "spinlock.h"
The common header "spinlock.h" contains spinlock implementation. It
may be used in LibOS, PAL, and other parts of Graphene. Previously,
it included "api.h" which is PAL-specific. This prevented using
"spinlock.h" in PAL-agnostic parts of Graphene such as mbedTLS
crypto adapters. This commit removes this dependency and fixes
a few emerged include-related bugs.