[Regression Test]: Fix /sys fs regression test
Current implementation incorrectly uses `_SC_ULONG_MAX`
to check the max ulong (instead of ULONG_MAX). This patch
addresses the issue.
Note: `SC_ULONG_MAX` is typically used with sysconf()
to inquire about the maximum value which can be stored
in a variable of type unsigned long int and is defined
to value of 117.

Signed-off-by: Vijay Dhanraj <vijay.dhanraj@intel.com>