[Pal] Fix {READ,WRITE}_ONCE macros
The old version of the macro (taken from Linux 5.6) seems to be buggy
(doesn't always emit an access) and the version from 5.11 almost works -
except for a GCC bug (https://gcc.gnu.org/bugzilla/show_bug.cgi?id=99258).
Fortunately it's not too hard to workaround it, and the new version
seems to work.

The bug was noticed because `sgx.preheat_enclave` didn't work on some
configurations.

After the changes some type mismatches started to get detected by the
compiler, so I also had to fix them.

Signed-off-by: Michał Kowalczyk <mkow@invisiblethingslab.com>