[LibOS/shim] rtld: Don't call __map_elf_object for internal ELFs
The __map_elf_object function had a special case for LibOS binary
(OBJECT_INTERNAL) and vDSO (OBJECT_VDSO) in which it only processed
already-mapped objects, effectively only recording information in
link_map structure.

This is now unnecessary, so removing this special case simplifies
the code: we now call __map_elf_object for ELFs that really need
to be mapped.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>