[LibOS/shim] rtld: Fix program header handling
The code (copied from glibc) intended to load first FILEBUF_SIZE
bytes, and then perform a separate load if necessary. However,
the comparison was actually in wrong direction, and a separate
load was performed only if it was *not* needed.

This change simplifies the code by always loading ELF header (Ehdr)
and program header (Phdr) separately.

In addition, we fix a memory leak in case when the program header
for the application (l_phdr) has to be allocated.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>