[LibOS] Add sync engine, and use it to synchronize file handles
This is an IPC subsystem that allows different processes exchange
data and lock access to resources. See shim_sync.h for the API and
information about usage.

As a proof of concept, the sync engine is used to synchronize
file handle information.

Because the sync engine is an experimental feature, it has to be
specially enabled in a manifest.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>