[GSC] Restrict NumPy and PyTorch examples to use only 8 threads
Previously, there was no restriction on the number of threads, so both
NumPy and PyTorch allocated as many threads as there are CPU cores on
the platform. On many-core platforms, this led to OOM errors in these
examples because enclave memory was quickly depleted (each thread in
Python allocates 128MB for its "memory arena").

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>