[Pal/Linux-SGX] EDMM batch allocation
SGX driver allocates EPC pages dynamically by faulting in
pages one at a time. This incurs a huge overhead due to
enclave exit for each page. To overcome this issue, a new
IOCTL has been introduced in the SGX driver which can take
the requested range and EAUG all the pages in one shot.
This patch updates graphene memory allocator to use the
newly added SGX driver IOCTL to pass range of EPC pages to
be allocated and on return simply EACCEPTs the EAUG'ed pages.

Signed-off-by: Vijay Dhanraj <vijay.dhanraj@intel.com>