[LibOS] Rework IPC transport layer
Old version was bugged, unstable, unnecessary complex and hard to
maintain.
After this PR other parts of code need not to worry about ipc
implementation details. As a bonus there is no 0.5 sleep on the program
exit.
IPC connections are now managed using two sets: incoming and outgoing.
The former is fully managed by a IPC worker thread, which handles new
connections and all incoming IPC messages. The latter is shared between
all threads, but encapsulated in the IPC module and the only thing you
need to send a IPC message is the destination process ID.

Signed-off-by: Borys Popławski <borysp@invisiblethingslab.com>