[Pal,LibOS] Make the logging system always flush data
This makes sure the behaviour of logging functions is consistent,
but that they still output whole lines with a single syscall
whenever possible.

The syscall trace (shim_parser.c), which depended on the buffering
mechanism in LibOS, is converted to use its own buffer. As a result,
we no longer need to keep a per-thread log buffer, just log prefix
for a given thread.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>