[LibOS] Get proper distlib path for python apps on Fedora and Ubuntu
The distlib path for python apps differs by distro like this:
- Ubuntu: /usr/lib/python3/dist-packages
- Fedora: /usr/lib64/python3.9/site-packages

This patch now builds these paths and only returns them if
the directory exists. The order in which we try things is important
because on Ubuntu /usr/lib/python3.6/site-packages may exist
but on Fedora /usr/lib/python3/dist-packages does not exist.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>