[Pal/Linux-SGX] Make AEX profiling code count as AEX handling
Previously if an async signal happened while inside AEX profiling code,
it was not considered as happening during AEX and as the result was
handled like it happened during an ocall. This caused the next host to
enclave transition to be EENTER (instead of ERESUME), which completely
desynchronized trusted<->untrusted states and either caused untrusted
part to crash or trusted part to kill the enclave.

Signed-off-by: Borys Popławski <borysp@invisiblethingslab.com>