[Pal/Linux-SGX] Fix vulnerability which allowed tampering with trusted files
Previously, Graphene generated `g_enclave_key` based on Intel SGX's
`EGETKEY(SEAL_KEY)`. This key was generated insecurely (it was always
the same for any enclave on the platform), and it was used to generate
hashes of trusted-file chunks. This was done via 128-bit AES-CMAC.
However, there is no need for a key-based hash to verify the integrity
of trusted-file chunks; SHA-256 (truncated to 128 bits) is sufficient.
Moreover, using AES-CMAC as a hashing function was a serious security
vulnerability: the attacker could re-generate the same key in her own
malicious enclave, and then make controlled changes to trusted files
without changing the hashes of trusted-file chunks. This would break
security of Graphene completely. Therefore, this commit removes
`g_enclave_key` and replaces its uses with keyless SHA-256.

Functions `load_trusted_file()` and `copy_and_verify_trusted_file()`
were slightly refactored as a side-effect of this change.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>