[Pal/Linux-SGX] Remove `g_enclave_key` and use SHA256 for file chunk hashes
Previously, Graphene generated `g_enclave_key` based on Intel SGX's
`EGETKEY(SEAL_KEY)`. This key was generated insecurely (it was always
the same for any enclave on the platform), and it was used to generate
hashes of trusted-file chunks. This was done via 128-bit AES-CMAC.
However, there is no need for a key-based hash to verify the integrity
of trusted-file chunks; SHA-256 (truncated to 128 bits) is sufficient.
Therefore, this commit removes `g_enclave_key` and replaces its uses
with keyless SHA-256.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>