[LibOS] Don't use `shim_mount.root`
The `root` field was intended to be an actual root of the filesystem,
but in practice, it was initialized completely outside of the dentry
tree (while the actual root was kept in the `mountpoint` field).

I'm planning to change mount semantics so that `root` is used again, but
for now, I'm removing the code that uses it, because it worked only by
accident and interferes with refactoring.

As a side-effect, this breaks the regression test that runs an
executable called "dev", because the /dev filesystem is mounted over it.
(Previously, the dentry for the executable was created under the
detached `shim_mount.root` dentry). I'm renaming the executable to
"devfs" instead, and adding a check that a dentry for executable is not
shadowed by a mount.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>