[Pal/Linux-SGX] Improve info and error messages in Protected Files
The `pf_crypt` utility now highlights that the encrypted-file path must
be specified exactly in the Graphene manifest (otherwise there will be a
mismatch in the metadata header of the encrypted file: the embedded
relative path will not match the `sgx.protected_files` path). Also,
error messages are printed now instead of raw error codes.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>