[LibOS] Simplify `getdents/getdents64`
This changes the filesystem `readdir` operation so that it only needs to
call a callback. This is a big simplification for underlying
filesystems: a lot of code was responsible for dynamically allocating
the right amount of memory.

Additionally, the only piece of information required from `readdir` is a
name, not file type or inode number. We perform a lookup on the
underlying dentries anyway, so this prevents code duplication.

The directory handle stores an array of dentries, along with size and
position in that array, instead of special `dot` and `dotdot` pointers.
This simplifies the implementation of `getdents/getdents64` system
calls, and also should allow for easier implementation of `lseek` on
directory handles.

Signed-off-by: Paweł Marczewski <pwmarcz@gmail.com>