[LibOS] Remove `shim_dentry.ino`
This commit removes the `shim_dentry.ino` field, and replaces it with a
function computing the dentry hash (which is what we usually did
anyway).

This is to simplify the code (the inode number was computed in several
places: `readdir`, `stat`, `hstat`), and to ensure that the inode number
is always available for a dentry.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>