[Common,Pal,LibOS] Replace `warn`/`__abort` with `log_always`/`abort`
The common library now expects that the environment specifies two
callbacks: `log_always` (previously called `warn`) and `abort`
(previously called `__abort`). The common library is built slightly
differently for different environments:
- For PAL environments (both trusted and untrusted), it expects
  `pal_log_always` and `pal_abort` symbols.
- For LibOS, it expects `shim_log_always` and `shim_abort` symbols.
- For all other environments (e.g., stand-alone SGX utilities, PAL
  regression tests, etc.) it expects `log_always` and `abort` symbols.

Strictly speaking, different callback names in LibOS, PAL and other
environments are not currently needed. But this is in preparation for
follow-up commits where LibOS and PAL will be statically linked
together (thus a need to avoid name collisions).

As a side effect, incorrect usages of `__abort()` are modified and some
includes are fixed.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>