[Common,Pal,LibOS] Move logging functionality to the common library
Logging macros are used in the common library, LibOS and PAL. Thus,
logging must be defined in the common library, with a callback defined
in the corresponding Graphene component. This enables the common library
functions to also use logging macros. This commit also removes unused
`pal_printf()`, `pal_vprintf()` and `debug()`.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>