[Pal] Add per-thread execution context (in-LibOS or in-PAL)
With this commit, at every moment in time each thread can be queried to
determine whether it executes in LibOS context or in PAL context. For
this, a 64-bit variable is introduced in the common PAL_TCB block. Each
bit in this variable denotes PAL context (bit == 0) or LibOS context
(bit == 1). To support nested exceptions, each of 64 bits represents a
nested level (bit 0 is normal execution, bit 1 is first exception level,
etc).

This per-thread execution context is not used currently. Future commits
will use it to dynamically choose between LibOS/PAL callbacks for
malloc/free/log/abort in the common library. This paves a way to
statically link LibOS, PAL and the common library in one binary.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>