[Pal/Linux-SGX] Force CPUID values of unused XSAVE features to zero
Previously, if the SGX enclave didn't use an XSAVE feature (by unsetting
the bit in the SECS.ATTRIBUTES.XFRM mask), the sanity check failed if
the XSAVE feature was available on the CPU. This check doesn't make
sense: the SGX enclave may want to "hide" some CPU features, but this
doesn't mean the enclave execution must be stopped. This commit removes
this incorrect check and instead forces the corresponding CPUID return
values to zero.

Additionally, this commit adds a missing check for extension offset.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>