[Pal] Fix sysfs NUMA sanitization when SNC is enabled
When sanitizing `/sys/devices/system/node/node0/distance`
path we compare the count of distances with number of nodes
present. But when SNC (Sub-Numa Clustering) mode is enabled,
we run out of buffer to store the distances which results in
incorrect distance count and the comparison failing. This
commit increases the buffer size to hold all NUMA distances
even for special cases like SNC.

Signed-off-by: Vijay Dhanraj <vijay.dhanraj@intel.com>