[Pal/Linux-SGX] Remove 32-bit limit on buffer size in read/write
Long time ago, there was a quirk that OCALLs used 32-bit sizes (e.g., in
`ocall_read`/`ocall_write`). This limitation was removed, but the checks
for uint32 overflow remained. This led e.g. to failures during fork,
when a checkpointed memory region exceeded 4GB and thus was deemed
"invalid" to be sent to the child.

This commit also adds the case "send >4GB checkpoint" to the
`large_mmap` regression test. However, since this takes several minutes
to execute on wimpy SGX machines, this case is commented out by default.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>