[LibOS,Pal] Emulate IPV6_V6ONLY both in LibOS and PAL
IPV6_V6ONLY socket flag is special in Graphene because it must be set
before listen/bind happens (in contrast to other flags that can be
applied after these syscalls). This led to special handling of
IPV6_V6ONLY in LibOS, and this attribute was completely emulated in
LibOS and not visible in PAL_STREAM_ATTR attributes. This led to corner
cases when `getsockopt(IPV6_V6ONLY)` tried to find this flag in
`pending_options` of the LibOS shim handle; however, pending options
could be already freed which led to a segfault.

This commit makes IPV6_V6ONLY less special: this flag is now known to
PAL and is handled similarly as other PAL_STREAM_ATTR attributes. This
change also fixes the segfault caused by `pending_options`
user-after-free bug. Also, surrounding code is slightly refactored.

Kudos to Denis Zygann (@dzygann) who found this bug and performed
initial analysis.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>