[LibOS] Do not reuse dentries after removing a file
This change ensures that dentries for removed files are not reused.
Instead of marking a dentry as negative, we mark it as invalid, and make
sure that a new one is created during next lookup.

This fixes a crash where we create a named pipe, remove it, and then
reuse the same object for a normal file. More generally, it ensures that
filesystems always receive a dentry in a consistent state.

Several unused dentry flags are also removed.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>