[Pal/Linux-SGX] Allocate program headers when reading ELFs
The code relied on the assumption that the program headers will fit
within FILEBUF_SIZE bytes from beginning of file. This is usually true
in practice, but is not guaranteed, and as a result Graphene could read
outside of the allocated buffer.

Instead, we perform two reads: first for the ELF header, then for the
program headers.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>