[LibOS] Remove FS_POLL_SZ
Now the intended way to check file handle size is to use `hstat`.

Removing FS_POLL_SZ uncovered a bug: in case of a handle for a file
deleted using `unlink`, `hstat` returned the right size, but
`poll(FS_POLL_SZ)` returned 0. When such a handle was mapped to memory,
LibOS did not checkpoint its contents, but recreated a mapping in the
child process.

Under Linux host, that allowed an application to create a shared mapping
for the same file in both processes, allowing them to communicate. This
feature was not intended (and not supported under Linux-SGX), but
unfortunately a few LTP tests depend on it. Now that such mappings
stopped working, these tests need to be skipped.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>