[Pal/Linux-SGX] Remove TODO on fetching SigRLs from IAS in RA-TLS
SigRLs are Signature Revocation Lists that can be separately fetched
from Intel Attestation Service (IAS), to verify that the EPID private
key of the user was revoked. Previously, we had a TODO comment to fetch
these SigRLs and verify against them. However, this is not needed
because the "verify attestation evidence" request to IAS (issued by
RA-TLS) already forces IAS to check the EPID key against SigRLs. Thus,
fetching SigRLs separately is just an optimization and not needed to be
enforced in RA-TLS.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>