[Pal/Linux-SGX] Suppress performing secret provisioning after execve
SGX Remote attestation and secret provisioning is performed via the
`LD_PRELOAD = "libsecret_prov_attest.so"` trick. Previously, there was a
bug in the `libsecret_prov_attest.so` library: its constructor was run
before the `main()` code in each execve'd process. In multi-process apps
that use execve syscall, this led to redundant remote attestation and
secret provisioning requests to the Secret provisioning service.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>