[Pal/Linux] Cap max available address for LibOS's memory management
During fork, the parent process checkpoints LibOS-managed memory regions
and the child process restores them at the exact same addresses.
However, the child process's main executable (libpal.so) may be mapped
randomly because we don't use fork for this emulation (which would
preserve the libpal.so memory mapping) but execve (which allows Linux
host to perform ASLR on libpal.so). This leads to transient failures of
the child because its libpal.so's memory region may overlap with
restored-from-checkpoint memory regions.

This commit restricts the memory range accessible to LibOS for memory
management. Since Linux allocates from the top of the x86-64 address
range (in approx. 0x7f... range), we hard-code a max available address
for LibOS memory management to definitely not overlap with 0x7f...
range.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>