[LibOS] Introduce inodes, rewrite `chroot` filesystem
This change begins the migration to inodes: file data will be kept in a
separate structure (inode) instead of dentry. In this step, inodes are
still optional, and only the `chroot` filesystem is rewritten to use
them.

The main use case for inodes is proper handling of `rename` and `unlink`
operations. The new `chroot` filesystem should do that much better than
the old one. I also added a regression test (`rename_unlink`) with
that checks various cases.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>